appease darglint!

+++ scrapli/driver/async_driver.py
@@ -132,7 +132,7 @@ class AsyncScrape(ScrapeBase):
             N/A
 
         Returns:
-            int: value of timeout_ops
+            float: value of timeout_ops
 
         Raises:
             N/A
